T3Board Constructor Example
In this first bit of code, I show you what a constructor is and why it's
important. I also show you the basic structure of writing your own class
and then using it in the main program. In this case, we do a simple
example with constructing a T3Board object that sets all of its
available play spaces to the "available" state.