[2] Implemented toString for T3Board
Now that we've covered objects and constructors, I added a String
function to the T3Board class so that we can retrieve a printable
representation of the game. This new code shows you how to write a
flexible method for a class. It is the primary way we will display the
game board to the players.