[3] Implemented playMove for T3Board
A new method has been implemented called playMove that will take an
integer position on the board to play as input and return a boolean
indicating the success of the operation. This method is tested in main
to ensure the program does not crash, even when bad input is given.